Added the ability to search through listings by title, and filter by price range. Added field price to Listing. Removed unused constructors from PropertyListingDTO and VehicleListingDTO.

+++ src/main/java/io/github/ksamodol/oglasnikbackend/controllers/LocationController.java
@@ -28,7 +28,7 @@ public class LocationController {
     }
 
     @GetMapping(path = "/place", params = "countyId")
-    public List<PlaceDTO> findAllPlacesByCountyId(@RequestParam int countyId){
+    public List<PlaceDTO> findAllPlacesByCountyId(@RequestParam Long countyId){
            return locationService.findAllPlacesByCountyId(countyId);
     }
